requirements.txt
Flask
torch
torchvision
Pillow
flask-cors
requests
